apps/close-relay/requirements.txt
flask==3.1.*
opentelemetry-api==1.29.*
opentelemetry-sdk==1.29.*
opentelemetry-exporter-otlp-proto-http==1.29.*
